requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.3.1
requests>=2.26.0
matplotlib>=3.4.3
tqdm>=4.62.3 